Actualizacion de codigo

+++ ControlDeMetas.Client/Contracts/IMetaClientService.cs
@@ -5,7 +5,7 @@ namespace ControlDeMetas.Client.Contracts
     public interface IMetaClientService
     {
         Task<List<Meta>> GetAll();
-        Task<Meta> GetById(int id);
+        Task<Meta> GetById(long id);
         Task Add(Meta meta);
         Task Update(int id, Meta meta);
         Task Delete(int id);
